requirements.txt
aiogram==3.18.0
langchain-gigachat==0.3.2
python-dotenv==1.0.0
httpx==0.27.0
uvicorn==0.29.0
starlette==0.37.2